+++ R/deprecated.R
@@ -30,6 +30,7 @@ split_by.data.frame <- split_by.tbl_ts
 #' @export
 split_by.tbl_df <- split_by.tbl_ts
 
+#' @description
 #' A thin wrapper of `dplyr::case_when()` if there are `NA`s
 #'
 #' @param formula A two-sided formula. The LHS expects a vector containing `NA`,
@@ -64,3 +65,27 @@ as.tsibble <- function(x, ...) {
 #' @keywords internal
 #' @include gaps.R
 fill_na <- fill_gaps.tbl_ts
+
+
+#' @description
+#' Find which row has duplicated key and index elements
+#'
+#' @rdname deprecated
+#' @param data A `tbl_ts` object.
+#' @param key Structural variable(s) that define unique time indices, used with
+#' the helper [id]. If a univariate time series (without an explicit key),
+#' simply call `id()`.
+#' @param index A bare (or unquoted) variable to specify the time index variable.
+#' @param fromLast `TRUE` does the duplication check from the last of identical
+#' elements.
+#'
+#' @return A logical vector of the same length as the row number of `data`
+#' @export
+find_duplicates <- function(data, key = id(), index, fromLast = FALSE) {
+  key <- use_id(data, !! enquo(key))
+  index <- validate_index(data, enquo(index))
+
+  grouped_df(data, vars = key) %>%
+    mutate(!! "zzz" := duplicated.default(!! index, fromLast = fromLast)) %>%
+    dplyr::pull(!! "zzz")
+}